User login added

+++ HallReservationSystem/app/Controller/UsersController.php
@@ -7,15 +7,16 @@ App::uses('AppController', 'Controller');
  */
 class UsersController extends AppController {
 
-/**
- * index method
- *
- * @return void
- */
-	public function index() {
-		$this->User->recursive = 0;
-		$this->set('users', $this->paginate());
-	}
+        public function beforeFilter() {
+            parent::beforeFilter();
+            $this->Auth->allow('add');
+        }
+
+
+        public function index() {
+                        $this->User->recursive = 0;
+                        $this->set('users', $this->paginate());
+        }
 
 /**
  * view method
@@ -28,8 +29,9 @@ class UsersController extends AppController {
 		if (!$this->User->exists($id)) {
 			throw new NotFoundException(__('Invalid user'));
 		}
-		$options = array('conditions' => array('User.' . $this->User->primaryKey => $id));
-		$this->set('user', $this->User->find('first', $options));
+//		$options = array('conditions' => array('User.' . $this->User->primaryKey => $id));
+//		$this->set('user', $this->User->find('first', $options));
+        $this->set('user', $this->User->read(null, $id));
 	}
 
 /**
@@ -38,15 +40,23 @@ class UsersController extends AppController {
  * @return void
  */
 	public function add() {
-		if ($this->request->is('post')) {
-			$this->User->create();
-			if ($this->User->save($this->request->data)) {
-				$this->Session->setFlash(__('The user has been saved'));
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('The user could not be saved. Please, try again.'));
-			}
-		}
+//		if ($this->request->is('post')) {
+//			$this->User->create();
+//			if ($this->User->save($this->request->data)) {
+//				$this->Session->setFlash(__('The user has been saved'));
+//				$this->redirect(array('action' => 'index'));
+//			} else {
+//				$this->Session->setFlash(__('The user could not be saved. Please, try again.'));
+//			}
+//		}
+            if($this->request->is('post')) {
+                $this->User->create();
+                if($this->User->save($this->request->data)) {
+                    $this->Session->setFlash(__('User has been saved'));
+                    return $this->redirect(array('action' => 'index'));
+                }
+                $this->Session->setFlash(__('User cannot be saved. Please try again'));
+            }
 	}
 
 /**
@@ -57,20 +67,35 @@ class UsersController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
-		if (!$this->User->exists($id)) {
-			throw new NotFoundException(__('Invalid user'));
-		}
-		if ($this->request->is('post') || $this->request->is('put')) {
-			if ($this->User->save($this->request->data)) {
-				$this->Session->setFlash(__('The user has been saved'));
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('The user could not be saved. Please, try again.'));
-			}
-		} else {
-			$options = array('conditions' => array('User.' . $this->User->primaryKey => $id));
-			$this->request->data = $this->User->find('first', $options);
-		}
+//		if (!$this->User->exists($id)) {
+//			throw new NotFoundException(__('Invalid user'));
+//		}
+//		if ($this->request->is('post') || $this->request->is('put')) {
+//			if ($this->User->save($this->request->data)) {
+//				$this->Session->setFlash(__('The user has been saved'));
+//				$this->redirect(array('action' => 'index'));
+//			} else {
+//				$this->Session->setFlash(__('The user could not be saved. Please, try again.'));
+//			}
+//		} else {
+//			$options = array('conditions' => array('User.' . $this->User->primaryKey => $id));
+//			$this->request->data = $this->User->find('first', $options);
+//		}
+            $this->User->id = $id;
+            if($this->User->exists()) {
+                throw new NotFoundException(__('Invalid User'));
+            }
+            if($this->request->is('post') || $this->request->is('put')) {
+                if($this->User->save($this->request->data)) {
+                    $this->Session->setFlash(__('This user has been saved'));
+                    return $this->redirect(array('action' => 'index'));
+                }
+                $this->Session->setFlash(__('This user cannot be saved. Please try again'));
+            }
+            else {
+                $this->request->data = $this->User->read(null, $id);
+                unset($this->request->data['User']['password']);
+            }
 	}
 
 /**
@@ -82,16 +107,33 @@ class UsersController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		$this->User->id = $id;
-		if (!$this->User->exists()) {
-			throw new NotFoundException(__('Invalid user'));
-		}
-		$this->request->onlyAllow('post', 'delete');
-		if ($this->User->delete()) {
-			$this->Session->setFlash(__('User deleted'));
-			$this->redirect(array('action' => 'index'));
-		}
-		$this->Session->setFlash(__('User was not deleted'));
-		$this->redirect(array('action' => 'index'));
+            if($this->request->is('post')) {
+                throw new MethodNotAllowedException();
+            }
+            
+            $this->User->id = $id;
+            if (!$this->User->exists()) {
+                    throw new NotFoundException(__('Invalid user'));
+            }
+            $this->request->onlyAllow('post', 'delete');
+            if ($this->User->delete()) {
+                    $this->Session->setFlash(__('User deleted'));
+                    $this->redirect(array('action' => 'index'));
+            }
+            $this->Session->setFlash(__('User was not deleted'));
+            $this->redirect(array('action' => 'index'));
 	}
+        
+        public function login() {
+            if($this->request->is('post')) {
+                if($this->Auth->login()) {
+                    return $this->redirect($this->Auth->redirect());
+                }
+                $this->Session->setFlash(__('Invalid Username or Password, try again!'));
+            }
+        }
+        
+        public function logout() {
+            return $this->redirect($this->Auth->logout());
+        }
 }
